Link GPIE01_01 100%
ONE HUNDRED
Final tweaks to two functions by me, then some help from Fable with some linking/weak function ordering issues that have been plaguing us for a while.
