+++ configure.py
@@ -352,7 +352,7 @@ config.libs = [
             Object(Matching, "jaudio/dspproc.c"),
             Object(Matching, "jaudio/dsptask.c"),
             Object(Matching, "jaudio/osdsp.c"),
-            Object(Equivalent, "jaudio/osdsp_task.c"),
+            Object(not MatchingFor("GPIP01_00"), "jaudio/osdsp_task.c"),
             Object(Matching, "jaudio/ipldec.c"),
             Object(Matching, "jaudio/dsp_cardunlock.c"),
             Object(Matching, "jaudio/driverinterface.c"),
@@ -495,7 +495,7 @@ config.libs = [
             Object(Matching, "plugPikiKando/interactEtc.cpp"),
             Object(Matching, "plugPikiKando/interactPullout.cpp"),
             Object(Matching, "plugPikiKando/saiEvents.cpp"),
-            Object(Equivalent, "plugPikiKando/simpleAI.cpp"),
+            Object(Matching, "plugPikiKando/simpleAI.cpp"),
             Object(Matching, "plugPikiKando/formationMgr.cpp"),
             Object(Matching, "plugPikiKando/globalShapes.cpp"),
             Object(Matching, "plugPikiKando/playerState.cpp"),
@@ -539,13 +539,13 @@ config.libs = [
             Object(Matching, "plugPikiKando/aiBore.cpp"),
             Object(Matching, "plugPikiKando/aiBoMake.cpp"),
             Object(Matching, "plugPikiKando/aiBou.cpp"),
-            Object(Equivalent, "plugPikiKando/aiBridge.cpp"),
+            Object(Matching, "plugPikiKando/aiBridge.cpp"),
             Object(Matching, "plugPikiKando/aiBreakWall.cpp"),
             Object(Matching, "plugPikiKando/aiTransport.cpp"),
             Object(Matching, "plugPikiKando/aiKinoko.cpp"),
             Object(Matching, "plugPikiKando/aiChase.cpp"),
             Object(Matching, "plugPikiKando/aiCrowd.cpp"),
-            Object(Equivalent, "plugPikiKando/aiDecoy.cpp"),
+            Object(Matching, "plugPikiKando/aiDecoy.cpp"),
             Object(Matching, "plugPikiKando/aiEnter.cpp"),
             Object(Matching, "plugPikiKando/aiEscape.cpp"),
             Object(Matching, "plugPikiKando/aiExit.cpp"),
@@ -636,7 +636,7 @@ config.libs = [
             Object(Matching, "plugPikiNakata/nlibgraphics.cpp"),
             Object(Matching, "plugPikiNakata/nlibmath.cpp"),
             Object(Matching, "plugPikiNakata/nlibspline.cpp"),
-            Object(Equivalent, "plugPikiNakata/nlibsystem.cpp"),
+            Object(Matching, "plugPikiNakata/nlibsystem.cpp"),
             Object(Matching, "plugPikiNakata/panianimator.cpp"),
             Object(Matching, "plugPikiNakata/panipikianimmgr.cpp"),
             Object(Matching, "plugPikiNakata/panitekianimator.cpp"),
@@ -823,7 +823,7 @@ config.libs = [
             Object(Matching, "plugPikiYamashita/ptclGenPack.cpp"),
             Object(Matching, "plugPikiYamashita/drawProgre.cpp"),
             Object(Matching, "plugPikiYamashita/spectrumCursorMgr.cpp"),
-            Object(Equivalent, "plugPikiYamashita/drawWorldMap.cpp"),
+            Object(Matching, "plugPikiYamashita/drawWorldMap.cpp"),
             Object(Matching, "plugPikiYamashita/drawCountDown.cpp"),
             Object(Matching, "plugPikiYamashita/drawGameOver.cpp"),
             Object(Matching, "plugPikiYamashita/yai.cpp"),
