+++ include/GemItem.h
@@ -2,6 +2,7 @@
 #define _GEMITEM_H
 
 #include "ItemMgr.h"
+#include "SAIEvent.h"
 #include "SimpleAI.h"
 #include "SoundMgr.h"
 #include "types.h"
