+++ include/nlib/Array.h
@@ -42,11 +42,15 @@ struct NArray {
 	}
 	virtual void set(int idx, T* elem) // _14
 	{
-		if (idx >= 0 && idx <= mMax - 1) {
-			mArray[idx] = elem;
-			if (idx + 1 > mCount) {
-				mCount = idx + 1;
-			}
+		if (idx < 0) {
+			return;
+		}
+		if (idx > mMax - 1) {
+			return;
+		}
+		mArray[idx] = elem;
+		if (idx + 1 > mCount) {
+			mCount = idx + 1;
 		}
 	}
 	virtual void add(T* elem) // _18
@@ -79,13 +83,17 @@ struct NArray {
 	}
 	virtual void remove(int idx) // _28
 	{
-		if (idx >= 0 && idx <= mCount - 1) {
-			for (int i = idx; i < mCount - 1; i++) {
-				mArray[i] = mArray[i + 1];
-			}
-
-			mCount--;
+		if (idx < 0) {
+			return;
 		}
+		if (idx > mCount - 1) {
+			return;
+		}
+		for (int i = idx; i < mCount - 1; i++) {
+			mArray[i] = mArray[i + 1];
+		}
+
+		mCount--;
 	}
 	virtual void add(int idx, T* elem) // _2C
 	{
@@ -97,13 +105,14 @@ struct NArray {
 	}
 	virtual void insert(int idx, T* elem) // _30
 	{
-		if (idx < mCount) {
-			for (int i = mCount; i >= idx + 1; i--) {
-				mArray[i] = mArray[i - 1];
-			}
-			mArray[idx] = elem;
-			mCount++;
+		if (idx >= mCount) {
+			return;
 		}
+		for (int i = mCount; i >= idx + 1; i--) {
+			mArray[i] = mArray[i - 1];
+		}
+		mArray[idx] = elem;
+		mCount++;
 	}
 	virtual T* get(int idx) // _34
 	{
